+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ## NEXT
 
+- Expose relatedSymbols from EnvironmentReduction. Needed for improving error
+  messages in LH
+  [#2346](https://github.com/ucsd-progsys/liquidhaskell/issues/2346).
+- Support extensionality in PLE [#704](https://github.com/ucsd-progsys/liquid-fixpoint/pull/704)
 
 ## 0.9.6.3.1 (2024-08-21)